Move all libloading stuff to a specific module
