+++ ash/src/lib.rs
@@ -23,13 +23,15 @@
 //!
 
 pub use crate::device::Device;
-#[cfg(feature = "libloading")]
-pub use crate::entry::{Entry, LoadingError};
 pub use crate::entry::{EntryCustom, InstanceError};
+#[cfg(feature = "libloading")]
+pub use crate::entry_libloading::{Entry, LoadingError};
 pub use crate::instance::Instance;
 
 mod device;
 mod entry;
+#[cfg(feature = "libloading")]
+mod entry_libloading;
 mod instance;
 pub mod prelude;
 pub mod util;
